GEMAH - T3.00 - E-Mails
- ADDED: Les emails sont maintenant présent sur la convention

+++ app/Http/Controllers/Scolarites/Documents/DecisionController.php
@@ -27,11 +27,11 @@ class DecisionController extends Controller
 	 */
 	private function generateFilename(Eleve $eleve, UploadedFile $file): string
 	{
-		$nom = $eleve->nom . '-' . $eleve->prenom;
+		$nom = $eleve->nom . "-" . $eleve->prenom;
 		$timestamp = Carbon::now()->timestamp;
 		$extension = $file->getClientOriginalExtension();
 
-		return $nom . '-' . $timestamp . '.' . $extension;
+		return $nom . "-" . $timestamp . "." . $extension;
 	}
 
 	/**
@@ -42,11 +42,11 @@ class DecisionController extends Controller
 	 */
 	private function stripAccents($str)
 	{
-		return strtr(utf8_decode($str), utf8_decode('àáâãäçèéêëìíîïñòóôõöùúûüýÿÀÁÂÃÄÇÈÉÊËÌÍÎÏÑÒÓÔÕÖÙÚÛÜÝ'), 'aaaaaceeeeiiiinooooouuuuyyAAAAACEEEEIIIINOOOOOUUUUY');
+		return strtr(utf8_decode($str), utf8_decode("àáâãäçèéêëìíîïñòóôõöùúûüýÿÀÁÂÃÄÇÈÉÊËÌÍÎÏÑÒÓÔÕÖÙÚÛÜÝ"), "aaaaaceeeeiiiinooooouuuuyyAAAAACEEEEIIIINOOOOOUUUUY");
 	}
 
 	/**
-	 * GET - Affiche le formulaire de création d'une Décision
+	 * GET - Affiche le formulaire de création d"une Décision
 	 *
 	 * @param Eleve $eleve
 	 * @return View
@@ -55,7 +55,7 @@ class DecisionController extends Controller
 	{
 		$enseignants = Enseignant::all();
 
-		return view('web.scolarites.eleves.documents.decisions.create', compact('eleve', 'enseignants'));
+		return view("web.scolarites.eleves.documents.decisions.create", compact("eleve", "enseignants"));
 	}
 
 	/**
@@ -67,56 +67,63 @@ class DecisionController extends Controller
 	 */
 	public function store(Request $request, Eleve $eleve): RedirectResponse
 	{
+		$dateBefore = Carbon::now()->addYear(25);
+		$dateAfter = Carbon::now()->subYear(25);
+
 		$request->validate([
-			'date_limite'       => 'nullable|date',
-			'date_cda'          => 'nullable|date',
-			'date_notification' => 'nullable|date',
-			'date_convention'   => 'nullable|date',
-			'numero_dossier'    => 'nullable|max:191',
-			'enseignant_id'     => 'nullable|integer',
-			'file'              => 'required',
+			"date_limite"       => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+			"date_cda"          => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+			"date_notification" => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+			"date_convention"   => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+			"numero_dossier"    => "nullable|max:191",
+			"enseignant_id"     => "nullable|exists:enseignants,id",
+			"file"              => "required",
 		]);
 
 		// On génère un nom de fichier
-		$filename = $this->generateFilename($eleve, $request->file('file'));
+		$filename = $this->generateFilename($eleve, $request->file("file"));
 
 		// On enregistre le fichier
-		$request->file('file')->storeAs('public/decisions/', $filename);
+		$request->file("file")->storeAs("public/decisions/", $filename);
 
 		$document = Document::create([
-			'nom'              => "Décision du " . Carbon::parse($request->input('date_notif'))->format('d/m/Y'),
-			'description'      => $request->input('description'),
-			'type_document_id' => TypeDocument::where('libelle', 'Décision')->first()->id,
-			'path'             => $filename,
-			'eleve_id'         => $eleve->id,
+			"nom"              => "Décision du " . Carbon::parse($request->input("date_notif"))->format("d/m/Y"),
+			"description"      => $request->input("description"),
+			"type_document_id" => TypeDocument::where("libelle", "Décision")->first()->id,
+			"path"             => $filename,
+			"eleve_id"         => $eleve->id,
 		]);
 
 		Decision::create([
-			'date_cda'          => $request->input('date_cda'),
-			'date_notification' => $request->input('date_notification'),
-			'date_limite'       => $request->input('date_limite'),
-			'date_convention'   => $request->input('date_convention'),
-			'numero_dossier'    => $request->input('numero_dossier'),
+			"date_cda"          => $request->input("date_cda"),
+			"date_notification" => $request->input("date_notification"),
+			"date_limite"       => $request->input("date_limite"),
+			"date_convention"   => $request->input("date_convention"),
+			"numero_dossier"    => $request->input("numero_dossier"),
 
-			'document_id'   => $document->id,
-			'enseignant_id' => $request->input('enseignant_id'),
+			"document_id"   => $document->id,
+			"enseignant_id" => $request->input("enseignant_id"),
 		]);
 
-		return redirect(route('web.scolarites.eleves.documents.index', [$eleve]));
+		return redirect(route("web.scolarites.eleves.documents.index", [$eleve]));
 	}
 
 	/**
-	 * GET - Affiche le formulaire d'édition d'une décision
+	 * GET - Affiche le formulaire d"édition d"une décision
 	 *
 	 * @param Eleve    $eleve
 	 * @param Decision $decision
-	 * @return \Illuminate\Http\Response
+	 * @return View|RedirectResponse
 	 */
 	public function edit(Eleve $eleve, Decision $decision)
 	{
-		$enseignants = Enseignant::all();
+		if ($decision->document->eleve_id == $eleve->id) {
+			$enseignants = Enseignant::all();
 
-		return view('web.scolarites.eleves.documents.decisions.edit', compact('eleve', 'decision', 'enseignants'));
+			return view("web.scolarites.eleves.documents.decisions.edit", compact("eleve", "decision", "enseignants"));
+		}
+
+		return back()->withErrors("Cette décision n'appartient pas à cet élève");
 	}
 
 	/**
@@ -129,41 +136,48 @@ class DecisionController extends Controller
 	 */
 	public function update(Request $request, Eleve $eleve, Decision $decision): RedirectResponse
 	{
-		$request->validate([
-			'date_limite'       => 'nullable|date',
-			'date_cda'          => 'nullable|date',
-			'date_notification' => 'nullable|date',
-			'date_convention'   => 'nullable|date',
-			'numero_dossier'    => 'nullable|max:191',
-			'enseignant_id'     => 'nullable|integer',
-		]);
+		$dateBefore = Carbon::now()->addYear(25);
+		$dateAfter = Carbon::now()->subYear(25);
 
-		if ($request->hasFile('file')) {
-			// On supprime l'ancien fichier
-			Storage::delete('public/' . $decision->document->path);
-
-			// On enregistre le fichier
-			$filename = $this->generateFilename($eleve, $request->file('file'));
-			$request->file('file')->storeAs('public/decisions/', $filename);
-
-			// Mise a jours du document
-			$decision->document->update([
-				'path' => $filename,
+		if ($decision->document->eleve_id == $eleve->id) {
+			$request->validate([
+				"date_limite"       => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+				"date_cda"          => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+				"date_notification" => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+				"date_convention"   => "nullable|date|before:{$dateBefore},after:{$dateAfter}",
+				"numero_dossier"    => "nullable|max:191",
+				"enseignant_id"     => "nullable|exists:enseignants,id",
 			]);
+
+			if ($request->hasFile("file")) {
+				// On supprime l"ancien fichier
+				Storage::delete("public/" . $decision->document->path);
+
+				// On enregistre le fichier
+				$filename = $this->generateFilename($eleve, $request->file("file"));
+				$request->file("file")->storeAs("public/decisions/", $filename);
+
+				// Mise a jours du document
+				$decision->document->update([
+					"path" => $filename,
+				]);
+			}
+
+			$decision->update([
+				"date_cda"          => $request->input("date_cda"),
+				"date_notification" => $request->input("date_notification"),
+				"date_limite"       => $request->input("date_limite"),
+				"date_convention"   => $request->input("date_convention"),
+				"numero_dossier"    => $request->input("numero_dossier"),
+
+				"enseignant_id" => $request->input("enseignant_id"),
+				"document_id"   => isset($document) ? $document->id : $decision->document_id,
+			]);
+
+			return redirect(route("web.scolarites.eleves.documents.index", [$eleve]));
 		}
 
-		$decision->update([
-			'date_cda'          => $request->input('date_cda'),
-			'date_notification' => $request->input('date_notification'),
-			'date_limite'       => $request->input('date_limite'),
-			'date_convention'   => $request->input('date_convention'),
-			'numero_dossier'    => $request->input('numero_dossier'),
-
-			'enseignant_id' => $request->input('enseignant_id'),
-			'document_id'   => isset($document) ? $document->id : $decision->document_id,
-		]);
-
-		return redirect(route('web.scolarites.eleves.documents.index', [$eleve]));
+		return back()->withErrors("Cette décision n'appartient pas à cet élève");
 	}
 
 	/**
@@ -178,7 +192,7 @@ class DecisionController extends Controller
 	{
 		if ($decision->document->eleve_id == $eleve->id) {
 			// On supprime le fichier
-			Storage::delete('public/decisions/' . $decision->document->path);
+			Storage::delete("public/decisions/" . $decision->document->path);
 
 			// On supprime la décision dans la BDD
 			$decision->delete();
@@ -202,7 +216,7 @@ class DecisionController extends Controller
 	public function download(Eleve $eleve, Decision $decision)
 	{
 		if ($decision->document->eleve_id == $eleve->id) {
-			return Storage::download('public/decisions/' . $decision->document->path, $this->stripAccents($decision->document->path));
+			return Storage::download("public/decisions/" . $decision->document->path, $this->stripAccents($decision->document->path));
 		}
 
 		return back()->withErrors("Cette decision n'appartient pas cet élève");
